% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{cb2connect}
\alias{cb2connect}
\title{connect to circBase2}
\usage{
cb2connect(host = "localhost", db.user = getOption("db.username"),
  password = rstudioapi::askForPassword("cb2 password"),
  dbname = getOption("mysql.db"))
}
\arguments{
\item{host}{hostname (localhost)}

\item{password}{cb2 password}

\item{dbname}{database name (circBase2)}

\item{user}{username (cb2admin)}
}
\description{
Connect to a local (default) or remote instance of circBase2.
}
\details{
A simple wrapper for DBI::dbConnect, exists pretty much only to
take advantage of password prompt.
}
