% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb2tools.R
\docType{package}
\name{cb2tools}
\alias{cb2tools}
\alias{cb2tools-package}
\title{cb2tools: circBase2 management tools}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

