% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{collapseVector}
\alias{collapseVector}
\title{collapse a character vector to SQL multiquery thing}
\usage{
collapseVector(charvec)
}
\arguments{
\item{charvec}{a character vector}
}
\description{
cook a character vector of baz, quux, quux, ... into format suitable for
SELECT * FROM foo WHERE bar IN ("baz", "quux", "quuux")
}
\details{
A simple wrapper for DBI::dbConnect, exists pretty much only to
take advantage of password prompt.
}
