% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{dbDump}
\alias{dbDump}
\title{load a ciRcus object into circBase2}
\usage{
dbDump(conn, CS, exp.annot)
}
\arguments{
\item{conn}{DBI connection}

\item{CS}{circRNAs (ciRcus object)}

\item{exp.annot}{experiment annotation (data table for now, would be nice to have a solid object down the line TODO)}
}
\description{
Do whatever it takes to push an object + annotation to circBase2
}
