% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getcircBaseIDs}
\alias{getcircBaseIDs}
\title{fetch circBase IDs from the database}
\usage{
getcircBaseIDs(DT)
}
\description{
for a given SummarizedExperiment, fetch circBase IDs based on chr:start-endStrand
for now, generate dummy temporary IDs
}
